tela de cada pokemon epecifico

+++ src/main/java/com/example/pokedex/entity/Pokemon.java
@@ -14,7 +14,7 @@ public class Pokemon implements Serializable{
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     
     private String name;
 
@@ -108,11 +108,11 @@ public class Pokemon implements Serializable{
         this.speed = speed;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
